ADD :: header - 헤더 부분과 메인 부분을 함께 보일 수 있도록 해냄
