+++ src/App.js
@@ -2,7 +2,6 @@ import { createGlobalStyle } from "styled-components";
 import styled from "styled-components";
 import Header from "./component/Header/index";
 import Main from "./component/Main/index";
-import Pointer from "./component/Pointer/index";
 
 const GlobalStyle = createGlobalStyle`
   body {
@@ -13,6 +12,10 @@ const GlobalStyle = createGlobalStyle`
 `;
 
 const Container = styled.div`
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: flex-start;
   width: 100vw;
   height: 100vh;
   background-color: #3e3d3f;
@@ -23,7 +26,6 @@ function App() {
     <>
       <GlobalStyle />
       <Container>
-        <Pointer />
         <Header />
         <Main />
       </Container>
